remove unecessary casts and fix templates

+++ TaxCalc/BusinessLogic/TaxCalculatorBL.cs
@@ -4,24 +4,24 @@ namespace TaxCalc.BusinessLogic
     public class TaxCalculatorBL : ITaxCalculatorBL
     {
         public TaxCalculatorBL() { }
-        public int CalculateTaxBracket(double grossIncome)
+        public double CalculateTaxBracket(double grossIncome)
         {
-            if (grossIncome <= 5000) return 0;
-            else if (grossIncome > 5000 && grossIncome <= 20000) return 20;
-            else if (grossIncome > 20000) return 40;
-            else return 0;
+            if (grossIncome <= 5000) return 0.0;
+            else if (grossIncome > 5000 && grossIncome <= 20000) return 20.0;
+            else if (grossIncome > 20000) return 40.0;
+            else return 0.0;
         }
 
         public double CalculateAnnualNetSalary(double grossIncome)
         {
             double val = grossIncome - 
-                (((double)CalculateTaxBracket(grossIncome) / 100) * grossIncome);
+                ((CalculateTaxBracket(grossIncome) / 100) * grossIncome);
             return Math.Round(val, 2);
         }
 
         public double CalculateAnnualTaxPaid(double grossIncome)
         {
-            double val = (((double)CalculateTaxBracket(grossIncome) / 100.0) * grossIncome);
+            double val = ((CalculateTaxBracket(grossIncome) / 100.0) * grossIncome);
             return Math.Round(val, 2);
         }
